Sound rework.
Moved sound code out of memc.c to seperate file. Sound is polled every 16
samples as determined by VIDC timings, rendered to a buffer at a 1-1.5 MHz
sample rate (depending on VIDC clock), has a 2.2 kHz second order low-pass
filter applied, then is downsampled to 48 kHz for output.

This fixes a bug with high sample rates that was present with the old code, and
provides a more authentically filtered experience.