Add Wild Vision MIDI Max emulation. Currently MIDI functionality has only been implemented for Windows.
As the podule configuration GUI hasn't been written yet, you'll have to manually
configure the MIDI in and out devices in the machine config file. An example
configuration for a podule in slot 3 would be :

[midimax.3]
midi_in_dev_nr=0
midi_out_dev_nr=0

Currently the available devices are listed in midimax_log.txt on startup.