Fix append_filename() to work reliably
sprintf() has undefined behaviour when destination and a source
both point to the same buffer.  This happens in podule-linux.c
and causes random failures to find the podule shared objects.

Using strdup() to create temporary copies of the source buffers
eliminates the undefined behaviour.

With this fixed it's now possible to use snprintf() instead
without segfaults, so do that.